// src/app/Models/usuarios.ts
export interface Usuarios {
  nombre: string;
  password: string;
  rol: string;
  apPaterno: string;
  apMaterno: string;
  edad: number;
  telefono: string;
  username: string;
  estado: string;
}
export interface UsuariosID {
  id: string;
  nombre: string;
  password: string;
  rol: string;
  apPaterno: string;
  apMaterno: string;
  edad: number;
  telefono: string;
  username: string;
  estado: string;
}